Change architecture of modules. Add Statistics module for the next stage of project.
