+++ src/KnBase/Structures.hs
@@ -1,10 +1,12 @@
-module KnBase.Types where
+module KnBase.Structures where
 
 import Prelude hiding (Word)
 import Text.Show.Unicode
 
 type Word = String
-type WordLemma = Word
+
+type Sentence = String
+type SentText = [Sentence]
 type Text = [Word]
 type MorphFile = String
 
@@ -25,6 +27,12 @@ instance Show Graph where
 data PartOfSpeech = Noun | Adjective | Verb | Adverb | Smth String
   deriving(Eq,Show)
 
+type WordWindow = [Word]
+type WordForms  = [Word]
+type WordLemma  = Word
+
+type WordInfo = (PartOfSpeech, WordForms, WordWindow)
+
 data WordProp = Prop
   { word :: Word
   , wLemma :: Word
@@ -35,35 +43,50 @@ data WordProp = Prop
 newtype AssocList k v = AssocList {alData :: [(k, v)]}
  deriving (Eq,Show)
 
-alAdd :: WordLemma -> PartOfSpeech -> Word 
+type Dictionary = AssocList WordLemma WordInfo
+
+
+alAdd :: WordLemma -> PartOfSpeech -> Word -> WordWindow
       -> Dictionary 
       -> Dictionary 
-alAdd wl pos w al1 = case lookup wl (alData al1) of
-  Nothing -> AssocList ((wl, (pos,[w])) : alData al1)
-  Just val -> AssocList (map change (alData al1))
+alAdd wl pos w window al1 = case lookup wl (alData al1) of
+  Nothing -> AssocList ((wl, (pos, [w], window)) : alData al1)
+  Just _ -> AssocList (map change (alData al1))
     where
-      change x@(xwl, (xpos, xws)) =
+      change x@(xwl, (xpos, xws, xwindow)) =
           if xwl == wl
           then
             if w `elem` xws
             then x
-            else (xwl, (xpos, w:xws))
+            else (xwl, (xpos, w:xws, window ++ xwindow))
           else x
 
 alFoldr :: ((k,a) -> b -> b) -> b -> AssocList k a -> b
 alFoldr f acc al = go acc al 
   where
-    go acc (AssocList []) = acc
-    go acc (AssocList (x:xs)) = f x (go acc (AssocList xs))
-
-type Dictionary = AssocList WordLemma (PartOfSpeech, [Word])
+    go accum (AssocList []) = accum
+    go accum (AssocList (x:xs)) = f x (go accum (AssocList xs))
 
 getDictWords :: WordLemma -> Dictionary -> [Word]
 getDictWords wl d = case lookup wl (alData d) of 
   Nothing -> ["NoSuchLemmaInDict"]
-  Just (_,ww) -> ww
+  Just (_,ww,_) -> ww
 
 getDictPOS:: WordLemma -> Dictionary -> PartOfSpeech 
 getDictPOS wl d = case lookup wl (alData d) of 
   Nothing -> Smth "NoSuchLemmaInDict"
-  Just (pos,_) -> pos
+  Just (pos,_,_) -> pos
+
+getDictWordWindow :: WordLemma -> Dictionary -> WordWindow
+getDictWordWindow wl d = case lookup wl (alData d) of 
+  Nothing -> ["NoSuchLemmaInDict"]
+  Just (_,_,window) -> window
+
+getDictLemma :: Word -> Dictionary -> Maybe WordLemma
+getDictLemma w d = findLemma (alData d)
+  where
+    findLemma :: [(WordLemma, WordInfo)] -> Maybe WordLemma
+    findLemma [] = Nothing 
+    findLemma ((wl,(_,wf,_)):xs) =
+      if w `elem` wf then Just wl
+                     else findLemma xs
